fix definitions in the net package
There were two issues: it was parsing all files regardless
of tags, and one of the _solaris.go files in net imported
a package which had all solaris tagged files, so no
available files, so import failed.

Also, the import name resolver needed a special case
for C files.

As a drive-by fix, we avoid a panic when trying to find the
definition of an identifier that's derived from a tuple
index assignment that has a type error.