Switching to using x/tools/go/packages
This also replaces the custom parser and type checker with the standard
ones.
Right now this is approxmately half the speed of the old one, but it
works and is correct in more cases, and works on top of modules.
There are some minor differences in output, things like files may be
absolute now, and constants are more likely to know their values etc.
Also it does not currently support passing an expression on the command
line.