Refactoring to allow for testing
This removes the fail function to normalise the return paths, and extracts the
actual jump to definition parts (as opposed to the flag handling and result
printing) into a testable function.
The goal is to add a collection of tests that should continue to pass when we
switch to the go/packages version.