more tests; some fixes
We add some more tests, including two failing tests, one
of which is fixed (it now doesn't ignore the in-editor version
of the buffer).

Also fix the go tools dependency so it refers to a working
version of the tools module that's not after the breaking
change made to the `contains:` pattern.

The issue that still needs to be fixed is that if you
jump to the definition of an embedded field from that
field itself, it should go to the type referred to rather
than going to the same place.

Still needed: a test that godef on an import path will
print the directory containing the imported package.
I'm not sure whether to stay within the packagestest
Expect framework for that one.