Fix crashes on arm linux.
Godef stared crashing again inside the vim-go plugin. It turned out to
be a simple dependency update. Lots of churn but it is all the result
of go mod vendor and go mod tidy.

With these updates godef is all happy again and we continue to be able
to avoid gopls that is entirely too slow for normal use.